First Commit
First Commit! I'm trying to remember the rules for fallthrough here. This isn't working quite right; what happens is when the first if statement runs, it also runs the else statement--but when the second statement runs, it ignores the else case.
I think it's that in this arrangement, two consecutive if statements followed by an else, the final if and the else statements are considered a pair.