if_else-if_else
This takes care of the extra printing of the integer after "Fizz" cases. However, this also precludes the printing of FizzBuzz for integers divisible by 15; reaching 15 will only produce "Fizz."
One obvious solution is to copy the second if-else statement and nest it inside the first statement. I'd like to think of a cleaner way to do it, though...