Building conway.java, still runtime issues
I'm running into an issue with running the file in order to test it--I strongly believe this is an ongoing issue with how I'm classifying and saving the files. I'm saving the work I've done here so that I can pull it back if need be, after I finish fiddling around again with my repository's organization.