Changed Tactics
Instead of reading in and working with plain strings as a means of recording ints, I decided to change my approach and work with ArrayList<Integer>s instead. Much more natural of an API, don't need to reinvent the wheel.
Still struggling with finding a solution for reading in the list of edges, but I'll have to work on it later.